+++ src/rcpp_textspace.cpp
@@ -94,6 +94,21 @@ Rcpp::List textspace_train(SEXP textspacemodel) {
   auto t_start = std::chrono::high_resolution_clock::now();
   Rcpp::Function format_posixct("format.POSIXct");
   Rcpp::Function sys_time("Sys.time");
+  
+  // Also collecting error at start before training (added with respect to Starspace code)
+  Rcpp::Rcout << Rcpp::as<std::string>(format_posixct(sys_time())) << " Initialising with learning rate " << rate << endl;
+  train_epoch.push_back(0);
+  train_rate.push_back(rate);
+  auto err = sp->model_->test(sp->trainData_, sp->args_->thread);
+  train_error.push_back(err);
+  Rcpp::Rcout << "                     > Training data error   " << err << endl;
+  if (sp->validData_ != nullptr) {
+    auto valid_err = sp->model_->test(sp->validData_, sp->args_->thread);
+    validation_error.push_back(valid_err);
+    Rcpp::Rcout << "                     > Validation data error " << valid_err << endl;
+  }
+  // Done collecting error at start before training (added with respect to Starspace code)
+  
   for (int i = 0; i < sp->args_->epoch; i++) {
     if (sp->args_->saveEveryEpoch && i > 0) {
       auto filename = sp->args_->model;
@@ -228,9 +243,12 @@ Rcpp::List textspace(std::string model = "textspace.bin",
     if(std::ifstream(basedoc))        args->basedoc = basedoc;
     if(std::ifstream(predictionFile)) args->predictionFile = predictionFile;
   }else if(load_from_r){
-    args->isTrain = true;  
+    args->isTrain = false;  
     //if(std::ifstream(initModel))      args->initModel = initModel;
-    if(std::ifstream(trainFile))      args->trainFile = trainFile;
+    if(std::ifstream(trainFile)){
+      args->isTrain = true;  
+      args->trainFile = trainFile; 
+    }
     if(std::ifstream(validationFile)) args->validationFile = validationFile;
   }else{
     Rcpp::stop("No valid trainFile nor testFile. Please check your path and check if the file is not opened.");
@@ -307,10 +325,22 @@ Rcpp::List textspace(std::string model = "textspace.bin",
       }
     }
     sp->initParser();
-    //sp->initDataHandler();
-    out = Rcpp::List::create(
-      Rcpp::Named("model") = sp,
-      Rcpp::Named("args") = textspace_args(sp));    
+    if(args->isTrain){
+      sp->parser_->resetDict(sp->dict_);
+      sp->initDataHandler();
+      Rcpp::List iter = textspace_train(sp);
+      if(save){
+        sp->saveModel(args->model);    
+      }
+      out = Rcpp::List::create(
+        Rcpp::Named("model") = sp,
+        Rcpp::Named("args") = textspace_args(sp),
+        Rcpp::Named("iter") = iter);
+    }else{
+      out = Rcpp::List::create(
+        Rcpp::Named("model") = sp,
+        Rcpp::Named("args") = textspace_args(sp)); 
+    }
   }else{
     if(args->isTrain){
       if(std::ifstream(args->initModel)){